requirement.txt
mysql-connector-python>=9.5.0
mypy>=1.18.1